allow only branch basics to publish on gh-page

+++ deploy-gh-pages.sh
@@ -6,4 +6,6 @@ git config user.name "SimonBaumannPro"
 git config user.email "simon.baumann@orange.com"
 git add .
 git commit -m "Deploy to GitHub Pages"
-git push --force --quiet "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git" master:gh-pages
+if [ "$TRAVIS_BRANCH" == "basics" ]; then
+	git push --force --quiet "https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git" master:gh-pages
+fi
